create an ApprovalManager class
