+++ app/models/ncr/work_order.rb
@@ -101,28 +101,9 @@ module Ncr
       self.approving_official && self.approving_official.email_address != approving_official_email
     end
 
-    # Check the approvers, accounting for frozen approving official
-    def approvers_emails
-      emails = self.system_approver_emails
-      if self.approver_email_frozen?
-        emails.unshift(self.approving_official.email_address)
-      else
-        emails.unshift(self.approving_official_email)
-      end
-      emails
-    end
-
     def setup_approvals_and_observers
-      emails = self.approvers_emails
-      if self.emergency
-        emails.each{|e| self.add_observer(e)}
-        # skip state machine
-        self.proposal.update(status: 'approved')
-      else
-        original_approvers = self.proposal.individual_approvals.non_pending.map(&:user)
-        self.force_approvers(emails)
-        self.notify_removed_approvers(original_approvers)
-      end
+      manager = ApprovalManager.new(self)
+      manager.setup_approvals_and_observers
     end
 
     def approving_official
@@ -302,24 +283,6 @@ module Ncr
     def self.update_comment_format(key, value, bullet, former=nil)
       from = former ? "from #{former} " : ''
       "#{bullet}*#{key}* was changed " + from + "to #{value}"
-    end
-
-    # Generally shouldn't be called directly as it doesn't account for
-    # emergencies, or notify removed approvers
-    def force_approvers(emails)
-      individuals = emails.map do |email|
-        user = User.for_email(email)
-        user.update!(client_slug: 'ncr')
-        # Reuse existing approvals, if present
-        self.proposal.existing_approval_for(user) || Approvals::Individual.new(user: user)
-      end
-      self.proposal.root_approval = Approvals::Serial.new(child_approvals: individuals)
-    end
-
-    def notify_removed_approvers(original_approvers)
-      current_approvers = self.proposal.individual_approvals.non_pending.map(&:user)
-      removed_approvers_to_notify = original_approvers - current_approvers
-      Dispatcher.on_approver_removal(self.proposal, removed_approvers_to_notify)
     end
   end
 end
